POSConstructorTypeErrors Fixed
